requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
deprecation==2.1.0
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pytest==8.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.7.0
python-multipart==0.0.9
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.37.2
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
python-keycloak==3.7.0
dlib==19.24.2
pytesseract==0.3.10
easyocr==1.7.1
google-generativeai==0.5.2
absl-py==2.1.0
albumentations==1.4.7
annotated-types==0.6.0
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.63.0
idna==3.7
imageio==2.34.1
intel-openmp==2021.4.0
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
Pillow==9.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycocotools==2.0.7
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.5.10
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
llama-index==0.10.18
llama-index-llms-groq==0.1.3
groq==0.4.2
llama-index-embeddings-huggingface==0.2.0